character: Relaxed the condition of State::RUNNING
When the character is running up/downwards a sloped surface,
the character will bump up and down, causing the body velocity.x
to be 0 for an instant. This will lead to a serious problem:
the character, while running, will repeatedly suddenly become idle
and then start running again, which looks really annoying.

To solve this problem, we can relax the condition in determineState()
by checking the body velocity.x from the previous box2d update().

Signed-off-by: Marco Wang <m.aesophor@gmail.com>